requirements.txt

Django>=3.0,<4.0
gunicorn==20.0.4
ptvsd==4.3.2
psycopg2-binary>=2.8
djangorestframework>=3.11.1,<4.0
djangorestframework_simplejwt==5.0.0
django-cors-headers>=3.5.0,<4.0
Pillow>=8.0,<9.0
requests>=2.0,<2.25.2
yagmail==0.14.256